fix: handle case where sell order id in sanity not available anymore
