+++ web-marketplace/src/components/templates/ProjectDetails/ProjectDetails.utils.tsx
@@ -22,6 +22,7 @@ import {
   AllCreditClassQuery,
   AllProjectPageQuery,
   PrefinanceTimelineItem,
+  Project as SanityProject,
 } from 'generated/sanity-graphql';
 import { UseStateSetter } from 'types/react/use-state';
 import { onBtnClick } from 'lib/button';
@@ -293,3 +294,22 @@ export const formatTimelineDates = (item: PrefinanceTimelineItem) =>
   `${formatDate(item.date, 'MMM YYYY')}${
     item.endDate ? ` - ${formatDate(item.endDate, 'MMM YYYY')}` : ''
   }`;
+
+export const getCardSellOrders = (
+  sanityFiatSellOrders: SanityProject['fiatSellOrders'],
+  sellOrders: UISellOrderInfo[],
+) =>
+  sanityFiatSellOrders
+    ?.map(fiatOrder => {
+      const sellOrder = sellOrders.filter(
+        cryptoOrder => cryptoOrder.id.toString() === fiatOrder?.sellOrderId,
+      )?.[0];
+      if (sellOrder) {
+        return {
+          ...fiatOrder,
+          ...sellOrder,
+        };
+      }
+      return null;
+    })
+    .filter(Boolean) || [];
